fix: resolve strict type issues flagged by Psalm 6

+++ lib/Support/PerformanceLoggerTask.php
@@ -46,8 +46,8 @@ class PerformanceLoggerTask {
 			$this->logger->debug(
 				sprintf(
 					$message . ' %d/%dMB memory used',
-					round(memory_get_usage() / 1024.0 / 1024.0),
-					round(memory_get_peak_usage() / 1024.0 / 1024.0)
+					round((float)memory_get_usage() / 1024.0 / 1024.0),
+					round((float)memory_get_peak_usage() / 1024.0 / 1024.0)
 				)
 			);
 		} else {
